+++ vendors/matomo/device-detector/Parser/OperatingSystem.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * Device Detector - The Universal Device Detection library for parsing User Agents
  *
@@ -10,7 +8,11 @@ declare(strict_types=1);
  * @license http://www.gnu.org/licenses/lgpl.html LGPL v3 or later
  */
 
+declare(strict_types=1);
+
 namespace DeviceDetector\Parser;
+
+use DeviceDetector\ClientHints;
 
 /**
  * Class OperatingSystem
@@ -40,18 +42,25 @@ class OperatingSystem extends AbstractParser
     protected static $operatingSystems = [
         'AIX' => 'AIX',
         'AND' => 'Android',
+        'ADR' => 'Android TV',
+        'AMZ' => 'Amazon Linux',
         'AMG' => 'AmigaOS',
-        'ATV' => 'Apple TV',
+        'ATV' => 'tvOS',
         'ARL' => 'Arch Linux',
         'BTR' => 'BackTrack',
         'SBA' => 'Bada',
         'BEO' => 'BeOS',
         'BLB' => 'BlackBerry OS',
         'QNX' => 'BlackBerry Tablet OS',
+        'BOS' => 'Bliss OS',
         'BMP' => 'Brew',
         'CAI' => 'Caixa Mágica',
         'CES' => 'CentOS',
+        'CST' => 'CentOS Stream',
+        'CLR' => 'ClearOS Mobile',
         'COS' => 'Chrome OS',
+        'CRS' => 'Chromium OS',
+        'CHN' => 'China OS',
         'CYN' => 'CyanogenMod',
         'DEB' => 'Debian',
         'DEE' => 'Deepin',
@@ -61,21 +70,31 @@ class OperatingSystem extends AbstractParser
         'FEN' => 'Fenix',
         'FOS' => 'Firefox OS',
         'FIR' => 'Fire OS',
+        'FOR' => 'Foresight Linux',
         'FRE' => 'Freebox',
         'BSD' => 'FreeBSD',
         'FYD' => 'FydeOS',
+        'FUC' => 'Fuchsia',
         'GNT' => 'Gentoo',
         'GRI' => 'GridOS',
         'GTV' => 'Google TV',
         'HPX' => 'HP-UX',
         'HAI' => 'Haiku OS',
+        'IPA' => 'iPadOS',
+        'HAR' => 'HarmonyOS',
         'HAS' => 'HasCodingOS',
         'IRI' => 'IRIX',
         'INF' => 'Inferno',
+        'JME' => 'Java ME',
         'KOS' => 'KaiOS',
+        'KAN' => 'Kanotix',
         'KNO' => 'Knoppix',
+        'KTV' => 'KreaTV',
         'KBT' => 'Kubuntu',
         'LIN' => 'GNU/Linux',
+        'LND' => 'LindowsOS',
+        'LNS' => 'Linspire',
+        'LEN' => 'Lineage OS',
         'LBT' => 'Lubuntu',
         'LOS' => 'Lumin OS',
         'VLN' => 'VectorLinux',
@@ -85,6 +104,7 @@ class OperatingSystem extends AbstractParser
         'MDR' => 'Mandriva',
         'SMG' => 'MeeGo',
         'MCD' => 'MocorDroid',
+        'MON' => 'moonOS',
         'MIN' => 'Mint',
         'MLD' => 'MildWild',
         'MOR' => 'MorphOS',
@@ -93,22 +113,33 @@ class OperatingSystem extends AbstractParser
         'MRE' => 'MRE',
         'WII' => 'Nintendo',
         'NDS' => 'Nintendo Mobile',
+        'NOV' => 'Nova',
         'OS2' => 'OS/2',
         'T64' => 'OSF1',
         'OBS' => 'OpenBSD',
+        'OWR' => 'OpenWrt',
+        'OTV' => 'Opera TV',
         'ORD' => 'Ordissimo',
+        'PAR' => 'Pardus',
         'PCL' => 'PCLinuxOS',
+        'PLA' => 'Plasma Mobile',
         'PSP' => 'PlayStation Portable',
         'PS3' => 'PlayStation',
+        'PUR' => 'PureOS',
         'RHT' => 'Red Hat',
+        'REV' => 'Revenge OS',
         'ROS' => 'RISC OS',
+        'ROK' => 'Roku OS',
         'RSO' => 'Rosa',
+        'ROU' => 'RouterOS',
         'REM' => 'Remix OS',
+        'REX' => 'REX',
         'RZD' => 'RazoDroiD',
         'SAB' => 'Sabayon',
         'SSE' => 'SUSE',
         'SAF' => 'Sailfish OS',
         'SEE' => 'SeewoOS',
+        'SIR' => 'Sirin OS',
         'SLW' => 'Slackware',
         'SOS' => 'Solaris',
         'SYL' => 'Syllable',
@@ -117,6 +148,7 @@ class OperatingSystem extends AbstractParser
         'S40' => 'Symbian OS Series 40',
         'S60' => 'Symbian OS Series 60',
         'SY3' => 'Symbian^3',
+        'TEN' => 'TencentOS',
         'TDX' => 'ThreadX',
         'TIZ' => 'Tizen',
         'TOS' => 'TmaxOS',
@@ -132,7 +164,8 @@ class OperatingSystem extends AbstractParser
         'WRT' => 'Windows RT',
         'XBX' => 'Xbox',
         'XBT' => 'Xubuntu',
-        'YNS' => 'YunOs',
+        'YNS' => 'YunOS',
+        'ZEN' => 'Zenwalk',
         'IOS' => 'iOS',
         'POS' => 'palmOS',
         'WOS' => 'webOS',
@@ -144,28 +177,32 @@ class OperatingSystem extends AbstractParser
      * @var array
      */
     protected static $osFamilies = [
-        'Android'               => ['AND', 'CYN', 'FIR', 'REM', 'RZD', 'MLD', 'MCD', 'YNS', 'GRI'],
+        'Android'               => [
+            'AND', 'CYN', 'FIR', 'REM', 'RZD', 'MLD', 'MCD', 'YNS', 'GRI', 'HAR',
+            'ADR', 'CLR', 'BOS', 'REV', 'LEN', 'SIR',
+        ],
         'AmigaOS'               => ['AMG', 'MOR'],
-        'Apple TV'              => ['ATV'],
         'BlackBerry'            => ['BLB', 'QNX'],
         'Brew'                  => ['BMP'],
         'BeOS'                  => ['BEO', 'HAI'],
-        'Chrome OS'             => ['COS', 'FYD', 'SEE'],
+        'Chrome OS'             => ['COS', 'CRS', 'FYD', 'SEE'],
         'Firefox OS'            => ['FOS', 'KOS'],
         'Gaming Console'        => ['WII', 'PS3'],
         'Google TV'             => ['GTV'],
         'IBM'                   => ['OS2'],
-        'iOS'                   => ['IOS', 'WAS'],
+        'iOS'                   => ['IOS', 'ATV', 'WAS', 'IPA'],
         'RISC OS'               => ['ROS'],
         'GNU/Linux'             => [
             'LIN', 'ARL', 'DEB', 'KNO', 'MIN', 'UBT', 'KBT', 'XBT', 'LBT', 'FED',
             'RHT', 'VLN', 'MDR', 'GNT', 'SAB', 'SLW', 'SSE', 'CES', 'BTR', 'SAF',
             'ORD', 'TOS', 'RSO', 'DEE', 'FRE', 'MAG', 'FEN', 'CAI', 'PCL', 'HAS',
-            'LOS', 'DVK',
+            'LOS', 'DVK', 'ROK', 'OWR', 'OTV', 'KTV', 'PUR', 'PLA', 'FUC', 'PAR',
+            'FOR', 'MON', 'KAN', 'ZEN', 'LND', 'LNS', 'CHN', 'AMZ', 'TEN', 'CST',
+            'NOV', 'ROU',
         ],
         'Mac'                   => ['MAC'],
         'Mobile Gaming Console' => ['PSP', 'NDS', 'XBX'],
-        'Real-time OS'          => ['MTK', 'TDX', 'MRE'],
+        'Real-time OS'          => ['MTK', 'TDX', 'MRE', 'JME', 'REX'],
         'Other Mobile'          => ['WOS', 'POS', 'SBA', 'TIZ', 'SMG', 'MAE'],
         'Symbian'               => ['SYM', 'SYS', 'SY3', 'S60', 'S40'],
         'Unix'                  => ['SOS', 'AIX', 'HPX', 'BSD', 'NBS', 'OBS', 'DFB', 'SYL', 'IRI', 'T64', 'INF'],
@@ -173,6 +210,25 @@ class OperatingSystem extends AbstractParser
         'Windows'               => ['WIN'],
         'Windows Mobile'        => ['WPH', 'WMO', 'WCE', 'WRT', 'WIO'],
         'Other Smart TV'        => ['WHS'],
+    ];
+
+    /**
+     * Contains a list of mappings from OS names we use to known client hint values
+     *
+     * @var array<string, array<string>>
+     */
+    protected static $clientHintMapping = [
+        'GNU/Linux' => ['Linux'],
+        'Mac'       => ['MacOS'],
+    ];
+
+    /**
+     * Operating system families that are known as desktop only
+     *
+     * @var array
+     */
+    protected static $desktopOsArray = [
+        'AmigaOS', 'IBM', 'GNU/Linux', 'Mac', 'Unix', 'Windows', 'BeOS', 'Chrome OS', 'Chromium OS',
     ];
 
     /**
@@ -196,25 +252,14 @@ class OperatingSystem extends AbstractParser
     }
 
     /**
-     * @inheritdoc
+     * Returns the os name and shot name
+     *
+     * @param string $name
+     *
+     * @return array
      */
-    public function parse(): ?array
+    public static function getShortOsData(string $name): array
     {
-        $return = $osRegex = $matches = [];
-
-        foreach ($this->getRegexes() as $osRegex) {
-            $matches = $this->matchUserAgent($osRegex['regex']);
-
-            if ($matches) {
-                break;
-            }
-        }
-
-        if (empty($matches)) {
-            return $return;
-        }
-
-        $name  = $this->buildByMatch($osRegex['name'], $matches);
         $short = 'UNK';
 
         foreach (self::$operatingSystems as $osShort => $osName) {
@@ -224,13 +269,79 @@ class OperatingSystem extends AbstractParser
 
             $name  = $osName;
             $short = $osShort;
+
+            break;
+        }
+
+        return \compact('short', 'name');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function parse(): ?array
+    {
+        $osFromClientHints = $this->parseOsFromClientHints();
+        $osFromUserAgent   = $this->parseOsFromUserAgent();
+
+        if (!empty($osFromClientHints['name'])) {
+            $name    = $osFromClientHints['name'];
+            $version = $osFromClientHints['version'];
+
+            // use version from user agent if non was provided in client hints, but os family from useragent matches
+            if (empty($version)
+                && self::getOsFamily($name) === self::getOsFamily($osFromUserAgent['name'])
+            ) {
+                $version = $osFromUserAgent['version'];
+            }
+
+            // If the OS name detected from client hints matches the OS family from user agent
+            // but the os name is another, we use the one from user agent, as it might be more detailed
+            if (self::getOsFamily($osFromUserAgent['name']) === $name && $osFromUserAgent['name'] !== $name) {
+                $name = $osFromUserAgent['name'];
+
+                if ('HarmonyOS' === $name) {
+                    $version = '';
+                }
+            }
+
+            $short = $osFromClientHints['short_name'];
+
+            // Chrome OS is in some cases reported as Linux in client hints, we fix this only if the version matches
+            if ('GNU/Linux' === $name
+                && 'Chrome OS' === $osFromUserAgent['name']
+                && $osFromClientHints['version'] === $osFromUserAgent['version']
+            ) {
+                $name  = $osFromUserAgent['name'];
+                $short = $osFromUserAgent['short_name'];
+            }
+        } elseif (!empty($osFromUserAgent['name'])) {
+            $name    = $osFromUserAgent['name'];
+            $version = $osFromUserAgent['version'];
+            $short   = $osFromUserAgent['short_name'];
+        } else {
+            return [];
+        }
+
+        $platform    = $this->parsePlatform();
+        $family      = self::getOsFamily($short);
+        $androidApps = ['com.hisense.odinbrowser', 'com.seraphic.openinet.pre', 'com.appssppa.idesktoppcbrowser'];
+
+        if (null !== $this->clientHints) {
+            if (\in_array($this->clientHints->getApp(), $androidApps) && 'Android' !== $name) {
+                $name    = 'Android';
+                $family  = 'Android';
+                $short   = 'ADR';
+                $version = '';
+            }
         }
 
         $return = [
             'name'       => $name,
             'short_name' => $short,
-            'version'    => $this->buildVersion((string) $osRegex['version'], $matches),
-            'platform'   => $this->parsePlatform(),
+            'version'    => $version,
+            'platform'   => $platform,
+            'family'     => $family,
         ];
 
         if (\in_array($return['name'], self::$operatingSystems)) {
@@ -263,6 +374,20 @@ class OperatingSystem extends AbstractParser
     }
 
     /**
+     * Returns true if OS is desktop
+     *
+     * @param string $osName OS short name
+     *
+     * @return bool
+     */
+    public static function isDesktopOs(string $osName): bool
+    {
+        $osFamily = self::getOsFamily($osName);
+
+        return \in_array($osFamily, self::$desktopOsArray);
+    }
+
+    /**
      * Returns the full name for the given short name
      *
      * @param string      $os
@@ -282,11 +407,140 @@ class OperatingSystem extends AbstractParser
     }
 
     /**
+     * Returns the OS that can be safely detected from client hints
+     *
+     * @return array
+     */
+    protected function parseOsFromClientHints(): array
+    {
+        $name = $version = $short = '';
+
+        if ($this->clientHints instanceof ClientHints && $this->clientHints->getOperatingSystem()) {
+            $hintName = $this->applyClientHintMapping($this->clientHints->getOperatingSystem());
+
+            foreach (self::$operatingSystems as $osShort => $osName) {
+                if ($this->fuzzyCompare($hintName, $osName)) {
+                    $name  = $osName;
+                    $short = $osShort;
+
+                    break;
+                }
+            }
+
+            $version = $this->clientHints->getOperatingSystemVersion();
+
+            if ('Windows' === $name) {
+                $majorVersion = (int) (\explode('.', $version, 1)[0] ?? '0');
+
+                if ($majorVersion > 0 && $majorVersion < 11) {
+                    $version = '10';
+                } elseif ($majorVersion > 10) {
+                    $version = '11';
+                }
+            }
+
+            if (0 === (int) $version) {
+                $version = '';
+            }
+        }
+
+        return [
+            'name'       => $name,
+            'short_name' => $short,
+            'version'    => $this->buildVersion($version, []),
+        ];
+    }
+
+    /**
+     * Returns the OS that can be detected from useragent
+     *
+     * @return array
+     *
+     * @throws \Exception
+     */
+    protected function parseOsFromUserAgent(): array
+    {
+        $osRegex = $matches = [];
+        $name    = $version = $short = '';
+
+        foreach ($this->getRegexes() as $osRegex) {
+            $matches = $this->matchUserAgent($osRegex['regex']);
+
+            if ($matches) {
+                break;
+            }
+        }
+
+        if (!empty($matches)) {
+            $name                                = $this->buildByMatch($osRegex['name'], $matches);
+            ['name' => $name, 'short' => $short] = self::getShortOsData($name);
+
+            $version = \array_key_exists('version', $osRegex)
+                ? $this->buildVersion((string) $osRegex['version'], $matches)
+                : '';
+
+            foreach ($osRegex['versions'] ?? [] as $regex) {
+                $matches = $this->matchUserAgent($regex['regex']);
+
+                if (!$matches) {
+                    continue;
+                }
+
+                if (\array_key_exists('name', $regex)) {
+                    $name                                = $this->buildByMatch($regex['name'], $matches);
+                    ['name' => $name, 'short' => $short] = self::getShortOsData($name);
+                }
+
+                if (\array_key_exists('version', $regex)) {
+                    $version = $this->buildVersion((string) $regex['version'], $matches);
+                }
+
+                break;
+            }
+        }
+
+        return [
+            'name'       => $name,
+            'short_name' => $short,
+            'version'    => $version,
+        ];
+    }
+
+    /**
+     * Parse current UserAgent string for the operating system platform
+     *
      * @return string
      */
     protected function parsePlatform(): string
     {
-        if ($this->matchUserAgent('arm|aarch64|Watch ?OS|Watch1,[12]')) {
+        // Use architecture from client hints if available
+        if ($this->clientHints instanceof ClientHints && $this->clientHints->getArchitecture()) {
+            $arch = \strtolower($this->clientHints->getArchitecture());
+
+            if (false !== \strpos($arch, 'arm')) {
+                return 'ARM';
+            }
+
+            if (false !== \strpos($arch, 'mips')) {
+                return 'MIPS';
+            }
+
+            if (false !== \strpos($arch, 'sh4')) {
+                return 'SuperH';
+            }
+
+            if (false !== \strpos($arch, 'x64')
+                || (false !== \strpos($arch, 'x86') && '64' === $this->clientHints->getBitness())
+            ) {
+                return 'x64';
+            }
+
+            if (false !== \strpos($arch, 'x86')) {
+                return 'x86';
+            }
+        }
+
+        if ($this->matchUserAgent('arm|aarch64|Apple ?TV|Watch ?OS|Watch1,[12]')) {
             return 'ARM';
         }
 
@@ -298,11 +552,11 @@ class OperatingSystem extends AbstractParser
             return 'SuperH';
         }
 
-        if ($this->matchUserAgent('WOW64|x64|win64|amd64|x86_?64')) {
+        if ($this->matchUserAgent('64-?bit|WOW64|(?:Intel)?x64|WINDOWS_64|win64|amd64|x86_?64')) {
             return 'x64';
         }
 
-        if ($this->matchUserAgent('(?:i[0-9]|x)86|i86pc')) {
+        if ($this->matchUserAgent('.+32bit|.+win32|(?:i[0-9]|x)86|i86pc')) {
             return 'x86';
         }